style(chat): clean MCP lint rules
