+++ apps/chat/lib/ai/mcp/mcp-oauth-provider.ts
@@ -20,21 +20,9 @@ import type { OAuthClientInformationFull } from "@/lib/db/mcp-queries";
 import type { McpOAuthSession } from "@/lib/db/schema";
 import { createModuleLogger } from "@/lib/logger";
 
+export { OAuthAuthorizationRequiredError } from "./oauth-authorization-required-error";
+
 const log = createModuleLogger("mcp-oauth-provider");
-
-/**
- * Custom error thrown when OAuth authorization is required.
- * The client should catch this and redirect the user to the authorization URL.
- */
-export class OAuthAuthorizationRequiredError extends Error {
-  authorizationUrl: URL;
-
-  constructor(authorizationUrl: URL) {
-    super("OAuth user authorization required");
-    this.name = "OAuthAuthorizationRequiredError";
-    this.authorizationUrl = authorizationUrl;
-  }
-}
 
 /**
  * PostgreSQL-backed OAuth client provider for MCP.
@@ -52,7 +40,8 @@ export class McpOAuthClientProvider implements OAuthClientProvider {
     serverUrl: string;
     clientMetadata: OAuthClientMetadata;
     onRedirectToAuthorization: (authUrl: URL) => Promise<void>;
-    state?: string; // Optional: adopt existing state (for callback reconciliation)
+    // Optional: adopt existing state for callback reconciliation.
+    state?: string;
   };
   private saveClientInformationPromise: Promise<void> | null = null;
 
@@ -61,7 +50,8 @@ export class McpOAuthClientProvider implements OAuthClientProvider {
     serverUrl: string;
     clientMetadata: OAuthClientMetadata;
     onRedirectToAuthorization: (authUrl: URL) => Promise<void>;
-    state?: string; // Optional: adopt existing state (for callback reconciliation)
+    // Optional: adopt existing state for callback reconciliation.
+    state?: string;
   }) {
     this.config = config;
   }
@@ -164,9 +154,9 @@ export class McpOAuthClientProvider implements OAuthClientProvider {
       ) {
         log.warn(
           {
-            state: authData.state,
-            savedRedirectUri: clientInfo.redirect_uris[0],
             currentRedirectUri: this.redirectUrl,
+            savedRedirectUri: clientInfo.redirect_uris[0],
+            state: authData.state,
           },
           "clientInformation: redirect URI mismatch, invalidating session"
         );
@@ -175,11 +165,10 @@ export class McpOAuthClientProvider implements OAuthClientProvider {
         }
         this.cachedAuthData = undefined;
         this.initialized = false;
-        return;
+      } else {
+        return clientInfo;
       }
-      return clientInfo;
     }
-    return;
   }
 
   async saveClientInformation(
@@ -204,16 +193,17 @@ export class McpOAuthClientProvider implements OAuthClientProvider {
       };
     }
 
-    this.saveClientInformationPromise = setOAuthClientInfoOnceByState({
-      state: this.currentOAuthState,
-      clientInfo: clientCredentials,
-    })
-      .then((session) => {
+    this.saveClientInformationPromise = (async () => {
+      try {
+        const session = await setOAuthClientInfoOnceByState({
+          clientInfo: clientCredentials,
+          state: this.currentOAuthState,
+        });
         this.cachedAuthData = session;
-      })
-      .finally(() => {
+      } finally {
         this.saveClientInformationPromise = null;
-      });
+      }
+    })();
     await this.saveClientInformationPromise;
   }
 
@@ -224,8 +214,8 @@ export class McpOAuthClientProvider implements OAuthClientProvider {
 
   async saveTokens(tokens: OAuthTokens): Promise<void> {
     this.cachedAuthData = await saveTokensAndCleanup({
-      state: this.currentOAuthState,
       mcpConnectorId: this.config.mcpConnectorId,
+      state: this.currentOAuthState,
       tokens,
     });
   }
@@ -258,9 +248,9 @@ export class McpOAuthClientProvider implements OAuthClientProvider {
     if (existingVerifier) {
       log.info(
         {
-          state: this.currentOAuthState,
           existingVerifierPrefix: existingVerifier.slice(0, 10),
           newVerifierPrefix: pkceVerifier.slice(0, 10),
+          state: this.currentOAuthState,
         },
         "saveCodeVerifier: SKIPPING - verifier already exists"
       );
@@ -270,8 +260,8 @@ export class McpOAuthClientProvider implements OAuthClientProvider {
 
     log.info(
       {
-        state: this.currentOAuthState,
         codeVerifierPrefix: pkceVerifier.slice(0, 10),
+        state: this.currentOAuthState,
       },
       "saveCodeVerifier: saving first verifier"
     );
@@ -285,16 +275,17 @@ export class McpOAuthClientProvider implements OAuthClientProvider {
     }
 
     // Serialize and make the DB write immutable (DB-side also guards against overwrite).
-    this.saveCodeVerifierPromise = setOAuthCodeVerifierOnceByState({
-      state: this.currentOAuthState,
-      codeVerifier: pkceVerifier,
-    })
-      .then((session) => {
+    this.saveCodeVerifierPromise = (async () => {
+      try {
+        const session = await setOAuthCodeVerifierOnceByState({
+          codeVerifier: pkceVerifier,
+          state: this.currentOAuthState,
+        });
         this.cachedAuthData = session;
-      })
-      .finally(() => {
+      } finally {
         this.saveCodeVerifierPromise = null;
-      });
+      }
+    })();
     await this.saveCodeVerifierPromise;
   }
 
@@ -302,9 +293,9 @@ export class McpOAuthClientProvider implements OAuthClientProvider {
     const authData = await this.getAuthData();
     log.info(
       {
-        state: this.currentOAuthState,
-        hasCodeVerifier: !!authData?.codeVerifier,
         codeVerifierPrefix: authData?.codeVerifier?.slice(0, 10),
+        hasCodeVerifier: !!authData?.codeVerifier,
+        state: this.currentOAuthState,
       },
       "codeVerifier called"
     );
@@ -338,9 +329,9 @@ export class McpOAuthClientProvider implements OAuthClientProvider {
     if (session.mcpConnectorId !== this.config.mcpConnectorId) {
       log.warn(
         {
-          state,
-          sessionConnectorId: session.mcpConnectorId,
           expectedConnectorId: this.config.mcpConnectorId,
+          sessionConnectorId: session.mcpConnectorId,
+          state,
         },
         "adoptState: connector ID mismatch"
       );
@@ -348,12 +339,12 @@ export class McpOAuthClientProvider implements OAuthClientProvider {
     }
     log.info(
       {
-        state,
-        previousState: this.currentOAuthState,
-        wasInitialized: this.initialized,
-        hasCodeVerifier: !!session.codeVerifier,
         hasClientInfo: !!session.clientInfo,
+        hasCodeVerifier: !!session.codeVerifier,
         hasTokens: !!session.tokens,
+        previousState: this.currentOAuthState,
+        state,
+        wasInitialized: this.initialized,
       },
       "adoptState: adopting session (overriding previous state if any)"
     );
